% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ffi_isquiet}
\alias{ffi_isquiet}
\title{Check if fluxfinder.quiet option is TRUE}
\usage{
ffi_isquiet()
}
\value{
TRUE or FALSE
}
\description{
Check if fluxfinder.quiet option is TRUE
}
\keyword{internal}
