% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ffi_normalize_time}
\alias{ffi_normalize_time}
\title{Normalize a vector of times}
\usage{
ffi_normalize_time(time, normalize = TRUE)
}
\arguments{
\item{time}{A vector of time values, either \code{\link{POSIXct}} or numeric}

\item{normalize}{Normalize the values so that first is zero? Logical}
}
\value{
A numeric vector of normalized values (if \code{normalize_time} is
TRUE) or the original vector if not.
}
\description{
Normalize a vector of times
}
\keyword{internal}
