% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{ffi_qaqc}
\alias{ffi_qaqc}
\title{Generate a QA/QC document}
\usage{
ffi_qaqc(
  flux_data,
  group_column,
  output_file = "qaqc.html",
  output_dir = getwd(),
  open_output = TRUE
)
}
\arguments{
\item{flux_data}{A data frame from \code{\link{ffi_compute_fluxes}} or similar}

\item{group_column}{Name of the grouping label column in \code{flux_data},
character; pass NULL to run with no grouping}

\item{output_file}{Name of the output file}

\item{output_dir}{Name of the output directory; default is current working directory}

\item{open_output}{Automatically open the output HTML file?}
}
\value{
The path of the output file
}
\description{
Generate a QA/QC document
}
\examples{
# Compute fluxes from the smart chamber concentration data
dat <- ffi_read_LIsmartchamber(f) # concentrations
dat$CO2_umol <- ffi_ppm_to_umol(dat$co2,
  volume = dat$TotalVolume[1] / 100 ^ 3,
  temp = dat$chamber_t[1])
dat$CO2_umol_m2 <- dat$CO2_umol /  0.0314 # normalize by area (20 cm collar)
dat$group <- paste(dat$label, dat$RepNum)
fluxes <- ffi_compute_fluxes(dat, group_column = "group",
  time_column = "TIMESTAMP", gas_column = "CO2_umol_m2", dead_band = 5)
x <- ffi_qaqc(fd, group_column = "Plot")
file.remove(x) # clean up
}
