% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ffi_read_LI7810}
\alias{ffi_read_LI7810}
\title{Read a LI-7810 data file}
\usage{
ffi_read_LI7810(file)
}
\arguments{
\item{file}{Filename to read, character}
}
\value{
A \code{\link{data.frame}} with the parsed data.
}
\description{
Read a LI-7810 data file
}
\details{
Currently LI-7810 and LI-7820 files are handled identically.
}
\examples{
f <- system.file("extdata/TG10-01087.data", package = "fluxfinder")
dat <- ffi_read_LI7810(f)
}
