% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ffi_read_LI78x0}
\alias{ffi_read_LI78x0}
\title{Internal utility function to read LI-78x0 files}
\usage{
ffi_read_LI78x0(file, model)
}
\arguments{
\item{file}{Filename to read, character}

\item{model}{Instrument model name, string}
}
\value{
A \code{\link{data.frame}} with the parsed data.
}
\description{
Internal utility function to read LI-78x0 files
}
\details{
The is an internal function used by \code{\link{ffi_read_LI7810}}
and \code{\link{ffi_read_LI7820}}, and not normally called by users.
}
\keyword{internal}
