% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc}
\alias{qaqc}
\title{Generate a QA/QC document}
\usage{
qaqc(output_file = "qaqc.html", output_dir = getwd())
}
\arguments{
\item{output_file}{Name of the output file}

\item{output_dir}{Name of the output directory; default is current working directory}
}
\value{
The path of the output file
}
\description{
Generate a QA/QC document
}
\examples{
x <- qaqc()
file.remove(x) # clean up
}
