% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wtf_compute_fluxes}
\alias{wtf_compute_fluxes}
\title{Compute fluxes for multiple groups (measurements)}
\usage{
wtf_compute_fluxes(
  data,
  group_column,
  time_column,
  conc_column,
  volume,
  area,
  dead_band = 0,
  normalize_time = TRUE,
  fit_function = wtf_fit_models,
  ...
)
}
\arguments{
\item{data}{A \code{\link{data.frame}} (or tibble or data.table)}

\item{group_column}{Name of the grouping column in \code{data}, character;
pass NULL to run with no grouping}

\item{time_column}{Name of the time column in \code{data}, character}

\item{conc_column}{Name of the gas concentration column in \code{data}, character}

\item{volume}{XXX}

\item{area}{XXX}

\item{dead_band}{Length of dead band, the equilibration period at the
beginning of the time series during which data are ignore, in seconds (numeric)}

\item{normalize_time}{Normalize the values so that first is zero? Logical}

\item{fit_function}{Optional flux-fit function;
default is \code{\link{wtf_fit_models}}}

\item{...}{Other parameters passed to \code{fit_function}}
}
\value{
A data.frame with one row per \code{group_column} value. It will
always include the mean value of \code{time_column} for that group, but other
columns depend on what is returned by the \code{fit_function}.
}
\description{
Compute fluxes for multiple groups (measurements)
}
\examples{
# No grouping
wtf_compute_fluxes(cars, group_column = NULL, "speed", "dist")
# With grouping
cars$Plot <- c("A", "B")
wtf_compute_fluxes(cars, "Plot", "speed", "dist")
# See the introductory vignette for a fully-worked example with real data
}
