% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{wtf_fill_metadata}
\alias{wtf_fill_metadata}
\title{Convenience function to fill in missing measurement metadata}
\usage{
wtf_fill_metadata(metadat, dead_band_default = 10, obs_length_default = 60)
}
\arguments{
\item{metadat}{Metadata, data frame}

\item{dead_band_default}{Default dead band value, numeric (seconds)}

\item{obs_length_default}{Default observation length value, numeric (seconds)}
}
\value{
The metadata data frame, with missing values filled in.
}
\description{
Convenience function to fill in missing measurement metadata
}
\examples{
metadat <- data.frame(Plot = 1:2, Dead_band = c(1, NA))
metadat <- wtf_fill_metadata(metadat)
# The missing entry is the 'Dead_band' column has been changed to the default
# (60), and a new column 'Obs_length' has been created and filled in
}
