% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wtf_fit_models}
\alias{wtf_fit_models}
\title{Fit various models to gas concentration data}
\usage{
wtf_fit_models(time, conc, quiet = FALSE)
}
\arguments{
\item{time}{Relative time of observation (typically seconds), numeric}

\item{conc}{Greenhouse gas concentration, numeric}

\item{quiet}{Be quiet? Logical}
}
\value{
A wide-form \code{\link{data.frame}} with fit statistics for linear,
robust linear, and polynomial models. By default, extensive details are
provided only for the linear fi; for robust linear and polynomial, only
the slope and R2 are returned, respectively, as these are intended for QA/QC.
}
\description{
Fit various models to gas concentration data
}
\details{
If a linear model cannot be fit, NULL is returned. If the robust
linear and/or polynomial models cannot be fit, then \code{NA} is returned
for their particular statistics.
}
\examples{
wtf_fit_models(cars$speed, cars$dist)
}
