% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_match.R
\name{wtf_metadata_match}
\alias{wtf_metadata_match}
\title{Match metadata info with a vector of data timestamps}
\usage{
wtf_metadata_match(
  data_timestamps,
  start_dates,
  start_times,
  obs_length_default = 60,
  obs_lengths = rep(obs_length_default, length(start_dates))
)
}
\arguments{
\item{data_timestamps}{Data timestamps, either character (YYYY-MM-DD HH:MM:SS) or \code{\link{POSIXct}}}

\item{start_dates}{Metadata measurement date entries, either character (YYYY-MM-DD) or \code{\link{POSIXct}}}

\item{start_times}{Metadata measurement start time entries, either character (HH:MM) or \code{\link[lubridate]{period}}}

\item{obs_length_default}{Default length of observations, in seconds (numeric)}

\item{obs_lengths}{Optional vector of custom observation lengths in seconds;
if provided must be the same length as \code{start_dates}.
Missing values will be filled in with the \code{obs_length_default} value}
}
\value{
A numeric vector equal in length to \code{data_timestamps}, with
each entry indicating the metadata entry that should be used for that
observation. \code{NA} is returned if a timestamp has no match in the metadata (i.e., does not
fall within any window defined by the \code{start_dates}, \code{start_times},
and observation length parameters).
}
\description{
Match metadata info with a vector of data timestamps
}
\examples{
# Data timestamps
d_t <- c("2024-01-01 13:00:05", "2024-01-01 13:00:10",
"2024-01-01 13:05:05", "2024-01-01 13:10:00")
# Metadata start dates and times: two measurements, starting 5 minutes apart
s_d <- c("2024-01-01", "2024-01-01")
s_t <- c("13:00", "13:05")
wtf_metadata_match(d_t, s_d, s_t)
# Returns {1, 1, 2, NA} indicating that the first and second data timestamps
# correspond to metadata entry 1, the third to entry 2, and the fourth
# has no match

# This generates an error because of overlapping timestamps:
\dontrun{
s_t <- c("13:00", "13:01")
wtf_metadata_match(d_t, s_d, s_t, obs_length_default = 120)
}
}
