% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wtf_ppm_to_umol}
\alias{wtf_ppm_to_umol}
\title{Convert ppm to micromoles using Ideal Gas Law}
\usage{
wtf_ppm_to_umol(ppm, volume, temp = 20, atm = 101325)
}
\arguments{
\item{ppm}{Gas concentration (ppmv), numeric}

\item{volume}{System volume (chamber + tubing + analyzer, m3), numeric}

\item{temp}{Optional ambient temperature (degrees C), numeric}

\item{atm}{Optional atmospheric pressure (Pa), numeric}
}
\value{
The value(s) in micromoles.
}
\description{
Convert ppm to micromoles using Ideal Gas Law
}
\note{
The defaults are NIST normal temperature and pressure.
}
\examples{
# 0.18 is the slope (ppm CO2/s) of the first observation in the
# TG10-01087 example data; see wtf_fit_models() examples
wtf_ppm_to_umol(0.18, volume = 0.1)
wtf_ppm_to_umol(0.18, volume = 0.1, temp = 5) # cold day
wtf_ppm_to_umol(0.18, volume = 0.1, temp = 5, atm = 75000) # in the Andes
}
\references{
Steduto et al.:
Automated closed-system canopy-chamber for continuous field-crop monitoring
of CO2 and H2O fluxes, Agric. For. Meteorol., 111:171-186, 2002.
\url{http://dx.doi.org/10.1016/S0168-1923(02)00023-0}
}
