% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{wtf_qaqc}
\alias{wtf_qaqc}
\title{Generate a QA/QC document}
\usage{
wtf_qaqc(output_file = "qaqc.html", output_dir = getwd(), open_output = TRUE)
}
\arguments{
\item{output_file}{Name of the output file}

\item{output_dir}{Name of the output directory; default is current working directory}

\item{open_output}{Automatically open the output HTML file?}
}
\value{
The path of the output file
}
\description{
Generate a QA/QC document
}
\examples{
x <- qaqc(open_output = FALSE)
file.remove(x) # clean up
}
