% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{wtf_read_LGR}
\alias{wtf_read_LGR}
\title{Read a LGR-xxxx data file}
\usage{
wtf_read_LGR(file, tz = "UTC")
}
\arguments{
\item{file}{Filename to read, character}

\item{tz}{Time zone of the file, character (optional)}
}
\value{
A \code{\link{data.frame}} with the parsed data.
}
\description{
Read a LGR-xxxx data file
}
\examples{
f <- system.file("extdata/LGR-data.csv", package = "whattheflux")
dat <- wtf_read_LGR(f)
}
