% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{wtf_read_LI7820}
\alias{wtf_read_LI7820}
\title{Read a LI-7820 data file}
\usage{
wtf_read_LI7820(file, quiet = FALSE)
}
\arguments{
\item{file}{Filename to read, character}

\item{quiet}{Be quiet? Logical}
}
\value{
A data frame with the parsed data.
}
\description{
Read a LI-7820 data file
}
\examples{
f <- system.file("extdata/TG20-01182.data", package = "whattheflux")
dat <- wtf_read_LI7820(f)
}
