% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{wtf_read_LI78x0}
\alias{wtf_read_LI78x0}
\title{Internal utility function to read LI-78x0 files}
\usage{
wtf_read_LI78x0(file, model, quiet)
}
\arguments{
\item{file}{Filename to read, character}

\item{model}{Model name, string}

\item{quiet}{Be quiet? Logical}
}
\value{
A data frame with the parsed data.
}
\description{
Internal utility function to read LI-78x0 files
}
