% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wtf_fit_models}
\alias{wtf_fit_models}
\title{Fit various models to gas concentration data}
\usage{
wtf_fit_models(time, conc, area, volume)
}
\arguments{
\item{time}{Relative time of observation (typically seconds), numeric}

\item{conc}{Greenhouse gas concentration (typically ppm or ppb), numeric}

\item{area}{Area covered by the measurement chamber (typically cm2), numeric}

\item{volume}{Volume of the system
(chamber + tubing + analyzer, typically cm3), numeric}
}
\value{
A wide-form \code{\link{data.frame}} with fit statistics for linear,
robust linear, and polynomial models. By default, extensive details are
provided only for the linear fi; for robust linear and polynomial, only
the slope and R2 are returned, respectively, as these are intended for QA/QC.
}
\description{
Fit various models to gas concentration data
}
\details{
If a linear model cannot be fit, NULL is returned. If the robust
linear and/or polynomial models cannot be fit, then \code{NA} is returned
for their particular statistics.
}
\examples{
# Toy data
wtf_fit_models(cars$speed, cars$dist)
# Real data
f <- system.file("extdata/TG10-01087.data", package = "whattheflux")
dat <- wtf_read_LI7810(f)[1:75,] # manually isolate first observation
dat$SECONDS <- dat$SECONDS - min(dat$SECONDS) # normalize time to start at 0
plot(dat$SECONDS, dat$CO2)
wtf_fit_models(dat$SECONDS, dat$CO2)
}
