% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{wtf_qaqc}
\alias{wtf_qaqc}
\title{Generate a QA/QC document}
\usage{
wtf_qaqc(
  flux_data,
  group_column,
  flux_column = "flux_estimate",
  int_column = "int_estimate",
  r2_column = "adj.r.squared",
  time1_column = "TIMESTAMP_min",
  time2_column = "TIMESTAMP_max",
  obs_data = NULL,
  obs_time_column = "TIMESTAMP",
  obs_gas_column = NULL,
  open_output = TRUE,
  ...
)
}
\arguments{
\item{flux_data}{A data frame from \code{\link{wtf_compute_fluxes}} or similar}

\item{group_column}{Name of the grouping label column in \code{flux_data},
character; pass NULL to run with no grouping}

\item{flux_column}{Name of the column with flux estimates; defaults to
"flux_estimate", the output of \code{\link{wtf_compute_fluxes}}}

\item{int_column}{Name of the column with intercept estimates; defaults to
"int_estimate", the output of \code{\link{wtf_compute_fluxes}}}

\item{r2_column}{Name of the column with adjusted R2 values; defaults to
"adj.r.squared", the output of \code{\link{wtf_compute_fluxes}}}

\item{time1_column}{time1_column}

\item{time2_column}{time2_column}

\item{obs_data}{Observational data, optional; a data.frame of concentration
data from which the \code{flux_data} were computed}

\item{obs_time_column}{obs_time_column}

\item{obs_gas_column}{obs_gas_column}

\item{open_output}{Automatically open the output HTML file?}

\item{...}{Other parameters passed on to \code{\link[rmarkdown]{render}}}
}
\value{
The path of the output file
}
\description{
Generate a QA/QC document
}
\examples{
# Toy data
cars$Plot <- LETTERS[1:5]
cars$TIMESTAMP <- cars$speed
fd <- wtf_compute_fluxes(cars, "Plot", "TIMESTAMP", "dist", dead_band = 5)
x <- wtf_qaqc(fd, group_column = "Plot")
file.remove(x) # clean up
# Pass in observations as well as fluxes for more complete QA/QC
x <- wtf_qaqc(fd, group_column = "Plot",
  obs_data = cars, obs_gas_column = "dist")
file.remove(x) # clean up
# See the introductory vignette for a fully-worked example with real data
}
