% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{wtf_qaqc}
\alias{wtf_qaqc}
\title{Generate a QA/QC document}
\usage{
wtf_qaqc(
  flux_data,
  group_column,
  flux_column = "slope_estimate",
  r2_column = "adj.r.squared",
  open_output = TRUE,
  ...
)
}
\arguments{
\item{flux_data}{A data frame from \code{\link{wtf_compute_fluxes}} or similar}

\item{group_column}{Name of the grouping label column in \code{flux_data},
character; pass NULL to run with no grouping}

\item{flux_column}{Name of the column with flux estimates; defaults to
"slope_estimate", the output of \code{\link{wtf_compute_fluxes}}}

\item{r2_column}{Name of the column with adjusted R2 values; defaults to
"adj.r.squared", the output of \code{\link{wtf_compute_fluxes}}}

\item{open_output}{Automatically open the output HTML file?}

\item{...}{Other parameters passed on to \code{\link[rmarkdown]{render}}}
}
\value{
The path of the output file
}
\description{
Generate a QA/QC document
}
\examples{
# Toy data
cars$Plot <- LETTERS[1:5]
fd <- wtf_compute_fluxes(cars, "Plot", "speed", "dist")
x <- wtf_qaqc(fd, group_column = "Plot")
file.remove(x) # clean up
# See the introductory vignette for a fully-worked example with real data
}
