% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{wtf_read_LI78x0}
\alias{wtf_read_LI78x0}
\title{Internal utility function to read LI-78x0 files}
\usage{
wtf_read_LI78x0(file, model, quiet)
}
\arguments{
\item{file}{Filename to read, character}

\item{model}{Instrument model name, string}

\item{quiet}{Be quiet? Logical}
}
\value{
A \code{\link{data.frame}} with the parsed data.
}
\description{
Internal utility function to read LI-78x0 files
}
\details{
The is an internal function used by \code{\link{wtf_read_LI7810}}
and \code{\link{wtf_read_LI7820}}, and not normally called by users.
}
